requirements.txt
asgiref==3.3.4
Django==3.2.4
django-colorfield==0.4.1
django-cors-headers==3.7.0
djangorestframework==3.12.4
mysql-connector==2.2.9
mysqlclient==2.0.3
pytz==2021.1
sqlparse==0.4.1
